Liver_Analysis/requirements.txt
numpy==1.23.5
opencv-python==4.5.5.64
tensorflow==2.10.0
matplotlib==3.5.2
pillow==9.2.0
scikit-learn==1.0.2
scikit-image==0.19.2
pandas==1.4.3
albumentations==1.2.1
scipy==1.11.4
PyWavelets==1.4.1 